test: split post action tests
